[build] Cleanup 64-bit compilation warnings
Fix integer and integer vs pointer size issues.

+++ datapipe.h
@@ -25,6 +25,7 @@
 # define _DATAPIPE_H_
 
 # include <stdbool.h>
+# include <stdint.h>
 # include <glib.h>
 
 /* ========================================================================= *
@@ -149,10 +150,10 @@ void  mce_datapipe_generate_inactivity (void);
  * ========================================================================= */
 
 /** Retrieve a gint from a datapipe */
-# define datapipe_get_gint(_datapipe) ((gint)(void*)datapipe_value(&(_datapipe)))
+# define datapipe_get_gint(_datapipe) ((gint)(intptr_t)(void*)datapipe_value(&(_datapipe)))
 
 /** Retrieve a guint from a datapipe */
-# define datapipe_get_guint(_datapipe) ((guint)(void*)datapipe_value(&(_datapipe)))
+# define datapipe_get_guint(_datapipe) ((guint)(uintptr_t)(void*)datapipe_value(&(_datapipe)))
 
 /* Helper for making display state requests
  *
